Added javascript modules, changed main js file to main.js from index.js
