+++ index.html
@@ -77,6 +77,6 @@
             <button id="change-tile">click me</button>
         </div>
         
-        <script src="index.js"></script>
+        <script type="module" src="main.js"></script>
     </body>
 </html>
